gtk: disable more printer stuff on non-unix platforms
Part of #232
