+++ gtk4/src/auto/functions.rs
@@ -10,6 +10,8 @@ use crate::AccessibleState;
 use crate::DebugFlags;
 use crate::PageSetup;
 use crate::PrintSettings;
+#[cfg(any(target_os = "linux", feature = "dox"))]
+#[cfg_attr(feature = "dox", doc(cfg(target_os = "linux")))]
 use crate::Printer;
 use crate::StyleContext;
 use crate::TextDirection;
@@ -154,6 +156,8 @@ pub fn disable_setlocale() {
 //    unsafe { TODO: call ffi:gtk_distribute_natural_allocation() }
 //}
 
+#[cfg(any(target_os = "linux", feature = "dox"))]
+#[cfg_attr(feature = "dox", doc(cfg(target_os = "linux")))]
 #[doc(alias = "gtk_enumerate_printers")]
 pub fn enumerate_printers<P: Fn(&Printer) -> bool + Send + Sync + 'static>(func: P, wait: bool) {
     assert_initialized_main_thread!();
